+++ plugins/record-hook/src/tests/objectTestData.ts
@@ -1,0 +1,26 @@
+import { Flatfile } from '@flatfile/api'
+
+export const defaultObjectValueSchema: Array<Flatfile.Property> = [
+  {
+    key: 'array',
+    type: 'enum',
+    config: {
+      options: [
+        {
+          value: 'defaultValue',
+          label: 'Default Value',
+        },
+        {
+          value: 'secondValue',
+          label: 'Second Value',
+        },
+      ],
+    },
+  },
+]
+
+export const defaultObjectValueData: Record<string, any>[] = [
+  {
+    array: 'defaultValue',
+  },
+]